feat(shared): PanZoomPattern component
