+++ sites/shared/components/mdx/tabbed-example.mjs
@@ -4,7 +4,7 @@ import { pluginBundle } from '@freesewing/plugin-bundle'
 import { pluginFlip } from '@freesewing/plugin-flip'
 import { pluginGore } from '@freesewing/plugin-gore'
 import { Design } from '@freesewing/core'
-import { Svg } from '../workbench/pattern/svg'
+import { Svg } from 'pkgs/react-components/src/index.mjs'
 import { Defs } from '../workbench/pattern/defs'
 import { Stack } from '../workbench/pattern/stack'
 import { useState, useEffect } from 'react'
