+++ sites/shared/components/workbench/index.mjs
@@ -38,7 +38,6 @@ export const Workbench = ({ design, Design, baseSettings, DynamicDocs, from }) =
   // Effect
   useEffect(() => {
     // Force re-render when baseSettings changes. Required when they are loaded async.
-    console.log('in effect')
     setSettings({ ...baseSettings, embed: true })
   }, [baseSettings])
